sync branch rt-smart.

+++ components/drivers/include/drivers/cputime.h
@@ -22,8 +22,8 @@ struct rt_clock_cputime_ops
 float    clock_cpu_getres(void);
 uint64_t clock_cpu_gettime(void);
 
-uint32_t clock_cpu_microsecond(uint32_t cpu_tick);
-uint32_t clock_cpu_millisecond(uint32_t cpu_tick);
+uint64_t clock_cpu_microsecond(uint64_t cpu_tick);
+uint64_t clock_cpu_millisecond(uint64_t cpu_tick);
 
 int clock_cpu_setops(const struct rt_clock_cputime_ops *ops);
